# config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Customer Management API",
    "version": "1.0.0",
    "description": "This API provides endpoints to manage customer data, including creating, updating, retrieving, and deleting customer information and associated orders.",
    "contact": {
      "name": "API Support",
      "email": "fiorezarn@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/customers": {
      "get": {
        "tags": ["Customer"],
        "summary": "Get All Customer",
        "description": "Get All Customer",
        "operationId": "getAllCustomer",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "customerId": {
                        "type": "integer",
                        "example": 1
                      },
                      "firstName": {
                        "type": "string",
                        "example": "Reza"
                      },
                      "lastName": {
                        "type": "string",
                        "example": "Naufal"
                      },
                      "gender": {
                        "type": "string",
                        "example": "male"
                      },
                      "birth": {
                        "type": "string",
                        "format": "date",
                        "example": "2002-02-19"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected server issue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unexpected server issue"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Customer"],
        "summary": "Create Customer",
        "description": "Create Customer",
        "operationId": "createCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "Reza"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Naufal"
                  },
                  "gender": {
                    "type": "string",
                    "example": "male"
                  },
                  "birth": {
                    "type": "string",
                    "format": "date",
                    "example": "2002-02-19"
                  }
                },
                "examples": {
                  "Create Customer": {
                    "value": {
                      "firstName": "Reza",
                      "lastName": "Naufal",
                      "gender": "male",
                      "birth": "2002-02-19"
                    }
                  },
                  "Missing Required Field": {
                    "value": {
                      "firstName": "Reza",
                      "gender": "male"
                    }
                  },
                  "Invalid Date Format": {
                    "value": {
                      "firstName": "Reza",
                      "gender": "male",
                      "birth": "19-02-2002"
                    }
                  },
                  "Unacceptable Gender Value": {
                    "value": {
                      "firstName": "Reza",
                      "gender": "unknown",
                      "birth": "2002-02-19"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer successfully created"
                    },
                    "customer": {
                      "type": "object",
                      "properties": {
                        "customerId": {
                          "type": "integer",
                          "example": 1
                        },
                        "firstName": {
                          "type": "string",
                          "example": "Reza"
                        },
                        "lastName": {
                          "type": "string",
                          "example": "Naufal"
                        },
                        "gender": {
                          "type": "string",
                          "example": "male"
                        },
                        "birth": {
                          "type": "string",
                          "format": "date",
                          "example": "2002-02-19"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Missing or invalid request body"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Customer already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Conflict"
                    },
                    "message": {
                      "type": "string",
                      "example": "Customer already exists"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unexpected issue"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "tags": ["Customer"],
        "summary": "Get Customer by ID",
        "description": "Retrieve a specific customer by their ID",
        "operationId": "getCustomerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the customer to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success - Customer found",
            "content": {
              "application/json": {
                "example": {
                  "message": "success get customer",
                  "data": {
                    "id": "12345",
                    "firstName": "reza",
                    "lastName": "Naufal",
                    "birth": "2002-02-19",
                    "gender": "male"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Customer"],
        "summary": "Update Customer",
        "description": "Update customer details by ID",
        "operationId": "updateCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the customer to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "birth": {
                    "type": "string",
                    "example": "2002-02-19"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "rafal"
                  },
                  "gender": {
                    "type": "string",
                    "example": "male"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "alqurni"
                  }
                }
              },
              "examples": {
                "Update Customer": {
                  "value": {
                    "birth": "2002-02-19",
                    "firstName": "rafal",
                    "gender": "male",
                    "lastName": "alqurni"
                  }
                },
                "Invalid Birth Date": {
                  "value": {
                    "birth": "02/19/2002",
                    "firstName": "rafal",
                    "gender": "male",
                    "lastName": "alqurni"
                  }
                },
                "Missing Last Name": {
                  "value": {
                    "birth": "2002-02-19",
                    "firstName": "rafal",
                    "gender": "male"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer successfully updated",
            "content": {
              "application/json": {
                "example": {
                  "message": "success update customer",
                  "data": {
                    "id": "12345",
                    "birth": "2002-02-19",
                    "firstName": "rafal",
                    "gender": "male",
                    "lastName": "alqurni"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid request body",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid input data"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Customer"],
        "summary": "Delete Customer",
        "description": "Delete a customer by ID",
        "operationId": "deleteCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the customer to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer successfully deleted",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer successfully deleted"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{id}/addresses": {
      "get": {
        "tags": ["Customer"],
        "summary": "Get Customer Address",
        "description": "Get the address of a specific customer by their ID",
        "operationId": "getCustomerAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the customer whose address is to be retrieved",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success - Customer address found",
            "content": {
              "application/json": {
                "example": {
                  "message": "success get customer address",
                  "data": {
                    "id": "12345",
                    "addresses": [
                      {
                        "street": "cattleya residence",
                        "city": "Tangerang Selatan",
                        "postalCode": "15418"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Customer or address not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer or address not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Customer"],
        "summary": "Create Address",
        "description": "Create an address for a specific customer",
        "operationId": "createAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the customer to create the address for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "example": "Tangerang Selatan"
                  },
                  "postalCode": {
                    "type": "string",
                    "example": "15418"
                  },
                  "street": {
                    "type": "string",
                    "example": "cattleya residence"
                  }
                }
              },
              "examples": {
                "Create Address": {
                  "value": {
                    "city": "Tangerang Selatan",
                    "postalCode": "15418",
                    "street": "cattleya residence"
                  }
                },
                "Missing City": {
                  "value": {
                    "postalCode": "15418",
                    "street": "cattleya residence"
                  }
                },
                "Invalid Postal Code": {
                  "value": {
                    "city": "Tangerang Selatan",
                    "postalCode": "abc123",
                    "street": "cattleya residence"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address successfully created",
            "content": {
              "application/json": {
                "example": {
                  "message": "success create address",
                  "data": {
                    "id": "12345",
                    "city": "Tangerang Selatan",
                    "postalCode": "15418",
                    "street": "cattleya residence"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid data",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid input data"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "customer not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected issue",
            "content": {
              "application/json": {
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}